requirements.txt
flask[async]==3.0.2
flask-cors==4.0.0
transformers==4.38.2
torch==2.2.1
python-dotenv==1.0.1
numpy==1.26.4 